Move `backend_call` from webapp and ingress into core
